pst-analysis-engine/api/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
boto3==1.41.0
opensearch-py==3.1.0
SQLAlchemy==2.0.44
alembic==1.17.2
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic-settings==2.12.0
email-validator==2.3.0
redis==7.1.0
celery==5.5.3
python-multipart==0.0.20
aiofiles==25.1.0
aiohttp==3.13.2  # For async HTTP client operations
requests==2.32.5
passlib[bcrypt]==1.7.4
python-jose==3.5.0
defusedxml==0.7.1  # Secure XML parsing
PyPDF2==3.0.1
reportlab==4.4.5
python-magic==0.4.27
# Note: pypff (libpff-python) for PST processing is optional and platform-specific
# Install separately if needed: pip install libpff-python
sentence-transformers==5.1.2
transformers==4.57.1
torch==2.9.1
numpy==2.3.5
requests-aws4auth==1.3.1

# General Utilities
pandas==2.3.3
pyyaml==6.0.3

# AI Model SDKs
anthropic>=0.39.0,<0.45.0  # Pin range for httpx compatibility
openai>=1.55.0
google-generativeai>=0.8.4
httpx>=0.27.0,<0.28.0  # Pin to 0.27.x - 0.28+ removed 'proxies' arg breaking anthropic

# Testing
pytest==9.0.1
pytest-asyncio==1.3.0
